// src/subscription-plans/interfaces/subscriptionPlan.ts
// Generated by https://quicktype.io

export interface SubscriptionPlan {
  _id: string;
  name: string;
  description: string;
  type: SubscriptionPlanType;
  price: number;
  durationMonths: number;
  stock: number;
  maxMenusPerRestaurant: number;
  maxImagesPerMenu: number;
  maxTablesPerRestaurant: number;
  maxUsersPerRestaurant: number;
  maxPerCategory: number;
  maxStoriesPerRestaurant: number;
  hasRealTimeOrders: boolean;
  hasUserManagement: boolean;
  hasRoleManagement: boolean;
  hasCategoryManagement: boolean;
  hasSalesManagement: boolean;
  hasReportsManagement: boolean;
  hasApiAccess: boolean;
  hasSettingsAccess: boolean;
  hasGoodSearchPosition: boolean;
  features: SubscriptionPlanFeature[];
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export interface SubscriptionPlanFeature {
  name: string;
  description: string;
  available: boolean;
  _id: string;
}

export enum SubscriptionPlanType {
  FREE = 'free',
  PREMIUM = 'premium',
  ENTERPRISE = 'enterprise',
  BASIC = 'basic',
}